requirements.txt
asgiref==3.4.1
backcall==0.2.0
black==21.11b1
click==8.0.3
decorator==5.1.0
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
flake8==4.0.1
ipython==7.30.1
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
